+++ sembast/lib/src/value_utils.dart
@@ -1,5 +1,11 @@
+@Deprecated('Use valuesAreEquals')
+
+/// Test where 2 values are equals, going deeper in lists and maps.
+bool valueAreEquals(Object? value1, Object? value2) =>
+    valuesAreEquals(value1, value2);
+
 /// Test where 2 values are equals, going deeper in lists and maps
-bool valueAreEquals(Object? value1, Object? value2) {
+bool valuesAreEquals(Object? value1, Object? value2) {
   if (value1 == null) {
     return (value2 == null);
   } else if (value2 == null) {
@@ -11,7 +17,7 @@ bool valueAreEquals(Object? value1, Object? value2) {
         return false;
       }
       for (var i = 0; i < value1.length; i++) {
-        if (!valueAreEquals(value1[i], value2[i])) {
+        if (!valuesAreEquals(value1[i], value2[i])) {
           return false;
         }
       }
@@ -24,7 +30,7 @@ bool valueAreEquals(Object? value1, Object? value2) {
         return false;
       }
       for (var key in value1.keys) {
-        if (!valueAreEquals(value1[key], value2[key])) {
+        if (!valuesAreEquals(value1[key], value2[key])) {
           return false;
         }
       }